feat: CreateProductRequest 타입 구현 #4
