+++ OpenMarket/OpenMarket/Network/APIRequest/APIRequest.swift
@@ -14,6 +14,7 @@ protocol APIRequest {
     var path: String { get }
     var method: String { get }
     var urlRequest: URLRequest? { get }
+    var header: [String: String]? { get }
     
 }
 
@@ -29,6 +30,10 @@ extension APIRequest {
     
     var urlRequest: URLRequest? {
         guard let url = URL(string: finalURL) else { return nil }
-        return URLRequest(url: url)
+        var request = URLRequest(url: url)
+        request.httpMethod = self.method
+        header?.forEach { request.addValue($1, forHTTPHeaderField: $0) }
+        return request
     }
+    
 }
