+++ OpenMarket/OpenMarket/Network/OpenMarketService.swift
@@ -25,13 +25,7 @@ extension OpenMarketService {
         case .checkHealth, .showProductPage, .showProductDetail:
             return makeURLRequest(url: url, header: [:])
         case .createProduct(let sellerID, let params, let images):
-            let boundary = UUID().uuidString
-            var request = makeURLRequest(url: url, header: [
-                "identifier": sellerID,
-                "Content-Type": "multipart/form-data; boundary=\(boundary)"
-            ])
-            request?.httpBody = makeMultiPartBody(params: params, images: images, boundary: boundary)
-            return request
+            
         case .updateProduct(let sellerID, _, let body):
             var request = makeURLRequest(url: url, header: [
                 "identifier": sellerID,
@@ -77,40 +71,8 @@ extension OpenMarketService {
 
 extension OpenMarketService {
     
-    private func makeURLRequest(url: URL, header: [String: String]) -> URLRequest? {
-        var request = URLRequest(url: url)
-        request.httpMethod = self.method
-        header.forEach { request.addValue($1, forHTTPHeaderField: $0) }
-        return request
-    }
     
-    private func configureBody(target: NSMutableData, name: String, data: Data, boundary: String) {
-        target.append("--\(boundary)\r\n")
-        target.append("Content-Disposition: form-data; name=\"\(name)\"\r\n")
-        target.append("Content-Type: application/json\r\n")
-        target.append("\r\n")
-        target.append(data)
-        target.append("\r\n")
-    }
-
-    private func configureBodyImage(target: NSMutableData, name:String, images: [Data], boundary: String) {
-        for image in images {
-            target.append("--\(boundary)\r\n")
-            target.append("Content-Disposition: form-data; name=\"\(name)\"; filename=\"\(UUID().uuidString).png\"\r\n")
-            target.append("Content-Type: image/png\r\n")
-            target.append("\r\n")
-            target.append(image)
-            target.append("\r\n")
-        }
-        
-    }
     
-    private func makeMultiPartBody(params: Data, images: [Data], boundary: String) -> Data {
-        let body = NSMutableData()
-        configureBody(target: body, name: "params", data: params, boundary: boundary)
-        configureBodyImage(target: body, name: "images", images: images, boundary: boundary)
-        body.append("--\(boundary)--\r\n")
-        return body as Data
-    }
+    
     
 }
